refactor(cli): consolidate version-fetching logic and remove unused files
This commit refactors the version-fetching logic by consolidating the eagerBackendVersion function into the whoami package, ensuring a single implementation is used across the CLI. Additionally, it removes obsolete files related to version handling, including version_eager.go, edge.go, and associated tests, streamlining the codebase and improving maintainability. The changes enhance the clarity of version management within the CLI commands.
